requirements.txt
PyYAML
jsonschema
python-dotenv
requests
psutil
discord.py==2.0.1
discord-ext-prometheus